feat: tweak styles
