+++ pages/marketplace/[organizerId]/[eventId]/[ticketId]/index.tsx
@@ -1,12 +1,4 @@
-import {
-  Alert,
-  Box,
-  Button,
-  Container,
-  IconButton,
-  Stack,
-  Typography
-} from "@mui/material"
+import { Container, IconButton, Stack, Typography } from "@mui/material"
 import { useRouter } from "next/router"
 import React, { useContext, useEffect, useState } from "react"
 import EventInfoBanner from "../../../../../components/market/event/EventInfoBanner"
@@ -17,7 +9,6 @@ import {
   EventTicket,
   TicketQuotaCheckResultDTO
 } from "../../../../../dtos/ticket/ticket.dto"
-import TicketCard from "../../../../../components/market/ticket/TicketCard"
 import axios from "axios"
 import {
   Event,
@@ -25,8 +16,6 @@ import {
 } from "../../../../../interfaces/event/event.interface"
 import { EventOrganizerUser } from "../../../../../interfaces/auth/user.interface"
 import { MarketTickets } from "../../../../../interfaces/market/marketEvent.interface"
-import { blue, green, grey, orange, purple } from "@mui/material/colors"
-import moment from "moment"
 import Head from "next/head"
 import TicketPurchaseModal from "../../../../../components/market/ticket/TicketPurchaseModal"
 import { useAccount } from "wagmi"
@@ -35,6 +24,9 @@ import Web3 from "web3"
 import { ABIItem } from "../../../../../interfaces/smartContract/smartContract.interface"
 import { UserContext } from "../../../../../contexts/user/UserContext"
 import { client } from "../../../../../configs/axios/axiosConfig"
+import { LayoutContext } from "../../../../../contexts/layout/LayoutContext"
+import TicketListingDetails from "../../../../../components/market/ticket/listing/TicketListingDetails"
+import TicketListingActions from "../../../../../components/market/ticket/listing/TicketListingActions"
 
 export const getServerSideProps: GetServerSideProps = async ({ params }) => {
   const eventId = params?.eventId
@@ -49,8 +41,6 @@ export const getServerSideProps: GetServerSideProps = async ({ params }) => {
         }
       }
     )
-
-    console.log(marketTicketRes)
 
     const { organizerInfo, event, ticketCollection } = marketTicketRes.data
     const tickets = [
@@ -84,14 +74,16 @@ interface PageProps {
 
 const MarketTicket = ({ ticket, event, organizer }: PageProps) => {
   const { address } = useAccount()
+  const { abi, web3 } = useSmartContract()
+  const { user } = useContext(UserContext)
+  const { setShowLoadingBackdrop } = useContext(LayoutContext)
   const router = useRouter()
+
   const [showPurchaseModal, setShowPurchaseModal] = useState(false)
   const [isOwnedTicket, setIsOwnedTicket] = useState(false)
   const [isResaleTicket, setIsResaleTicket] = useState(false)
   const [isSold, setIsSold] = useState(false)
   const [statusChecked, setStatusChecked] = useState(false)
-  const { abi, web3 } = useSmartContract()
-  const { user } = useContext(UserContext)
   const [ticketQuotaCheckResult, setTicketQuotaCheckResult] =
     useState<TicketQuotaCheckResultDTO>()
 
@@ -99,7 +91,6 @@ const MarketTicket = ({ ticket, event, organizer }: PageProps) => {
     return (
       // This ticket has been sold before
       ticket.ownerHistory.length > 1 &&
-      ticket.ownerHistory.at(-1) !== ticket.ownerHistory.at(0) &&
       // The user is not the ticket's most recent owner
       ticket.ownerHistory.at(-1) !== address &&
       // Ticket is marked as for sale in the smart contract
@@ -113,6 +104,9 @@ const MarketTicket = ({ ticket, event, organizer }: PageProps) => {
     address: string,
     force?: boolean
   ) => {
+    // Show loading backdrop
+    setShowLoadingBackdrop(true)
+
     // Check the user's ticket purchase quota
     const _ticketQuotaCheckResult = await client.get("/ticket/quota/check", {
       params: {
@@ -126,7 +120,6 @@ const MarketTicket = ({ ticket, event, organizer }: PageProps) => {
     // If the ticket hasn't been bought yet (no smartContractTicketId), no need to check for ownership
     const smartContractTicketId = ticket.smartContractTicketId
     if (!force && smartContractTicketId === undefined) {
-      setStatusChecked(true)
       return
     }
 
@@ -138,19 +131,19 @@ const MarketTicket = ({ ticket, event, organizer }: PageProps) => {
       .then((result: any) => {
         setIsOwnedTicket(result.owner === address)
         setIsResaleTicket(checkResaleTicket(result.forSale))
-        setStatusChecked(true)
+        setIsSold(
+          ticket.ownerHistory.length > 1 &&
+            ticket.ownerHistory.at(-1) !== address &&
+            !isResaleTicket
+        )
+        setTimeout(() => {
+          setStatusChecked(true)
+        }, 1000)
       })
       .catch(() => setStatusChecked(true))
-  }
-
-  function getEventLocationUrl() {
-    if (event.online_url) {
-      return event.online_url
-    } else if (event.location?.place_id) {
-      return `https://www.google.com/maps/place/?q=place_id:${event.location.place_id}`
-    }
-
-    return "/marketplace"
+      .finally(() => {
+        setShowLoadingBackdrop(false)
+      })
   }
 
   // Checks ticket ownership
@@ -158,10 +151,7 @@ const MarketTicket = ({ ticket, event, organizer }: PageProps) => {
     // If the dependencies aren't ready, don't do anything
     if (!web3 || !abi || !event.smartContractAddress || !address) return
     checkTicketOwnership(web3, abi, address)
-    setIsSold(
-      ticket.ownerHistory.length > 1 && ticket.ownerHistory.at(-1) !== address
-    )
-  }, [web3, abi, address])
+  }, [web3, abi, address, isResaleTicket])
 
   return (
     <>
@@ -176,13 +166,13 @@ const MarketTicket = ({ ticket, event, organizer }: PageProps) => {
           setOpen={setShowPurchaseModal}
         />
       )}
-      <Container>
-        <BannerLayout
-          backgroundImage={event.image as string}
-          title=""
-          subtitle=""
-          actionName=""
-        >
+      <BannerLayout
+        backgroundImage={event.image as string}
+        title=""
+        subtitle=""
+        actionName=""
+      >
+        <Container>
           <EventInfoBanner
             event={event}
             imgFallbackSrc={organizer.profileImage as string}
@@ -206,207 +196,25 @@ const MarketTicket = ({ ticket, event, organizer }: PageProps) => {
             </IconButton>
             <Typography>Back to all tickets</Typography>
           </Stack>
-          <Box
-            sx={{
-              border: 2,
-              borderRadius: 2,
-              padding: 2,
-              backgroundColor: "white"
-            }}
-          >
-            <Stack direction="row" sx={{ justifyContent: "space-between" }}>
-              <Box sx={{ position: "relative" }}>
-                <TicketCard
-                  ticketId={ticket?._id ?? ""}
-                  image={ticket?.ticketMetadata[0].image as string}
-                  name={ticket?.name ?? ""}
-                  ticketTypeLabel={ticket?.ticketType ?? ""}
-                  price={(ticket?.price?.amount ?? "").toString()}
-                  sx={
-                    isSold
-                      ? { filter: "saturate(0.5)", opacity: 0.7 }
-                      : undefined
-                  }
-                />
-                {isSold && (
-                  <Typography
-                    variant="h3"
-                    fontWeight="600"
-                    color={grey[100]}
-                    sx={{ position: "absolute", top: "80px", left: "40px" }}
-                  >
-                    SOLD
-                  </Typography>
-                )}
-              </Box>
-              <Box sx={{ width: "100%", marginLeft: 4 }}>
-                {isResaleTicket && (
-                  <Alert
-                    severity="info"
-                    sx={{
-                      backgroundColor: blue[500],
-                      color: "white",
-                      marginBottom: 2
-                    }}
-                  >
-                    This is a resale ticket by #addressId
-                  </Alert>
-                )}
-                <Stack spacing={2}>
-                  <Typography variant="h6" component="h1">
-                    Ticket Details
-                  </Typography>
-                  <Stack
-                    direction="row"
-                    justifyContent="space-between"
-                    alignItems="center"
-                  >
-                    <Typography color={grey[500]}>Ticket Name</Typography>
-                    <Typography>{ticket.name}</Typography>
-                  </Stack>
-                  <Stack
-                    direction="row"
-                    justifyContent="space-between"
-                    alignItems="center"
-                  >
-                    <Typography color={grey[500]}>Description</Typography>
-                    <Typography>{ticket.description}</Typography>
-                  </Stack>
-                  <Stack
-                    direction="row"
-                    justifyContent="space-between"
-                    alignItems="center"
-                  >
-                    <Typography color={grey[500]}>Type</Typography>
-                    <Typography>{ticket.ticketType}</Typography>
-                  </Stack>
-                  <Typography variant="h6" component="h1">
-                    Event Details
-                  </Typography>
-                  <Stack
-                    direction="row"
-                    justifyContent="space-between"
-                    alignItems="center"
-                  >
-                    <Typography color={grey[500]}>Event Date</Typography>
-                    <Typography>
-                      {moment(event.startDate).format("lll")}
-                    </Typography>
-                  </Stack>
-                  <Stack
-                    direction="row"
-                    justifyContent="space-between"
-                    alignItems="center"
-                  >
-                    <Typography color={grey[500]}>Door Time</Typography>
-                    <Typography>
-                      {moment(event.doorTime).format("lll")}
-                    </Typography>
-                  </Stack>
-                  <Stack
-                    direction="row"
-                    justifyContent="space-between"
-                    alignItems="center"
-                  >
-                    <Typography color={grey[500]}>Location</Typography>
-                    <a
-                      target="_blank"
-                      href={getEventLocationUrl()}
-                      rel="noopener noreferrer"
-                    >
-                      <Typography>
-                        {event.location?.structured_formatting.main_text ||
-                          event.online_url}
-                      </Typography>
-                    </a>
-                  </Stack>
-                </Stack>
-              </Box>
-            </Stack>
-          </Box>
-          {isOwnedTicket && (
-            <Alert
-              action={
-                <Button
-                  color="inherit"
-                  size="small"
-                  onClick={() => router.push(`/me/${ticket._id}`)}
-                >
-                  View Ticket
-                </Button>
-              }
-              sx={{ border: `2px solid ${green[100]}`, mt: 2 }}
-            >
-              You are the owner of this ticket
-            </Alert>
-          )}
-          {statusChecked && !isOwnedTicket && !isSold && (
-            <Stack
-              direction="row"
-              justifyContent="space-between"
-              sx={{
-                marginY: 4,
-                border: 2,
-                borderRadius: 2,
-                padding: 2,
-                backgroundColor: "white"
-              }}
-            >
-              <Stack alignItems="start">
-                <Typography>Want to claim this ticket? Buy now</Typography>
-                <Typography fontWeight="700">
-                  {ticket.price.amount} {ticket.price.currency}
-                </Typography>
-              </Stack>
-              {user ? (
-                <Button
-                  variant="contained"
-                  disabled={!ticketQuotaCheckResult?.allowPurchase}
-                  onClick={() => {
-                    setShowPurchaseModal(true)
-                  }}
-                  sx={{
-                    background: purple[400],
-                    "&:hover": {
-                      background: purple[500]
-                    }
-                  }}
-                >
-                  <Typography fontWeight={600}>Purchase Ticket</Typography>
-                </Button>
-              ) : (
-                <Button
-                  variant="contained"
-                  onClick={() => {
-                    router.push({
-                      pathname: "/auth/login",
-                      query: {
-                        referrer: window.location.pathname
-                      }
-                    })
-                  }}
-                >
-                  <Typography>Log in to purchase</Typography>
-                </Button>
-              )}
-            </Stack>
-          )}
-          {statusChecked && !ticketQuotaCheckResult?.allowPurchase && (
-            <Alert
-              severity="warning"
-              sx={{
-                mt: 2,
-                border: `2px solid ${orange[100]}`
-              }}
-            >
-              You have purchased{" "}
-              {ticketQuotaCheckResult?.ownedTicketsCount.toString()} out of{" "}
-              {ticketQuotaCheckResult?.quota} tickets per person. You can not
-              purchase any more tickets from this event.
-            </Alert>
-          )}
-        </BannerLayout>
-      </Container>
+          <TicketListingDetails
+            ticket={ticket}
+            event={event}
+            statusChecked={statusChecked}
+            isSold={isSold}
+            isResaleTicket={isResaleTicket}
+          />
+          <TicketListingActions
+            user={user}
+            ticket={ticket}
+            statusChecked={statusChecked}
+            isSold={isSold}
+            isResaleTicket={isResaleTicket}
+            isOwnedTicket={isOwnedTicket}
+            ticketQuotaCheckResult={ticketQuotaCheckResult}
+            setShowPurchaseModal={setShowPurchaseModal}
+          />
+        </Container>
+      </BannerLayout>
     </>
   )
 }
